+++ code/prefabs/candy_fruit.lua
@@ -33,7 +33,12 @@ local function fn(Sim)
     inst.components.edible.foodtype = "VEGGIE"
     inst.components.edible.healthvalue = 0
     inst.components.edible.hungervalue = 5
-    inst.components.edible.sanityvalue = 5
+    inst.components.edible.sanityvalue = 10
+
+    inst:AddComponent("perishable")
+    inst.components.perishable:SetPerishTime(600)
+    inst.components.perishable:StartPerishing()
+    inst.components.perishable.onperishreplacement = "spoiled_food" 
 
     inst:AddComponent("fuel")
     inst.components.fuel.fuelvalue = 5